Refactor scoring classes to clarify their roles
The code now uses
- LocalScore to compute simple local scores for a node given parents,
- CandidateRestrictedScore to compute the local score sums
  constrained by candidate parents,
- CandidateComplementScore to compute the local score sums
  complementary to those constrained by candidate parents
  and constrained by maximum indegree d,
- Score to compute the sum of the two previous sums.

The Score object is the one which the MCMC finally polls for scores.
DAGR needs to use both CandidateRestrictedScore and
CandidateComplementScore.

Finally, there is also score_array (with a reference to in a number)
of places, which is a 2d array for storing the local scores
constrained by the candidate parents.