Update profiling results, profiler code
time 6. Precompute DAGR now takes longer, as the f table is copied to
Python from C++ (and uses double memory as its not deleted in C++).

time 7. Sample DAGS did not decrease as hoped, so will have to roll-back or
implement everything in C++. The values for the 10 runs vary a lot:
92.092, 54.436, 205.705, 177.915, 260.436, 215.783, 9.081, 70.279,
201.031, 125.305.

mem 3. Not sure what caused this increase.