Sampling parameters from the BGe prior. (#17)
Sample parameters from the BGe prior and posterior in the Gaussian case

Also added sampling of the mean. The expected neighbourhood size was lowered to 2 for the test_causal.py, so that the test would go through. In some cases the MSE can be substantially high and it is pure luck if one gets the test through. Minor error correction the MSE calculation in test_causal.